app/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pandas
requests
dotenv
